+++ src/sys/dev/pcmcia/if_ep_pcmcia.c
@@ -1,3 +1,4 @@
+/**	$MirOS$ */
 /*	$OpenBSD: if_ep_pcmcia.c,v 1.31 2004/05/12 06:35:11 tedu Exp $	*/
 /*	$NetBSD: if_ep_pcmcia.c,v 1.16 1998/08/17 23:20:40 thorpej Exp $  */
 
@@ -336,7 +337,7 @@ ep_pcmcia_attach(parent, self, aux)
 		return;
 	}
 
-	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, psc->sc_pcioh.size);
+	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, (int)(psc->sc_pcioh.size));
 
 	switch (pa->product) {
 	case PCMCIA_PRODUCT_3COM_3C562:
